moved AddressInfo and related types from explorer to dbtypes
